Planai perziura sukurimas

+++ database/migrations/2018_01_30_143041_create_plans_fees_table.php
@@ -16,9 +16,9 @@ class CreatePlansFeesTable extends Migration
         Schema::create('plans_fees', function (Blueprint $table) {
             $table->increments('id');
             $table->integer('plan_id');
-            $table->integer('price_sms');
-            $table->integer('price_calls');
-            $table->integer('price_gb');
+            $table->double('price_sms');
+            $table->double('price_calls');
+            $table->double('price_gb');
 
         });
     }
